feat(approval): implement AI workflow pipeline

+++ backend/src/Services/ApprovalService/Domain/Models/ClassificationResult.cs
@@ -3,7 +3,7 @@
     public class ClassificationResult
     {
         public string SuggestedCategory { get; set; } = string.Empty;
-        public double Confidence { get; set; }
+        public decimal Confidence { get; set; }
         public string Reasoning { get; set; } = string.Empty;
         public List<string> DetectedViolations { get; set; } = new List<string>();
     }
